requirements.txt
streamlit
numpy
opencv-python
Pillow
